signal/signal: creates package for monitoring of OS signals
